Se aumenta el tamaño de los tipos de datos para los identificadores de cliente, dirección, pedido del POS y teléfono. También se añade el tipo de entrada en OrderToStore para saber si fueron migrados o no

+++ ConnectCallCenter/Hubs/ClientHub.cs
@@ -108,7 +108,7 @@ namespace ConnectCallCenter.Hubs
         }
 
         [HubMethodName(SharedConstants.Server.CALCULATE_RECURRENCE_CLIENT_HUB_METHOD), UsedImplicitly]
-        public ResponseMessageData<RecurrenceResponseModel> CalculateRecurrence(List<int> lstClientId)
+        public ResponseMessageData<RecurrenceResponseModel> CalculateRecurrence(List<long> lstClientId)
         {
             try
             {
