+++ Drs.Model/Address/AddressPhoneModel.cs
@@ -2,9 +2,9 @@
 {
     public class AddressPhoneModel
     {
-        public int AddressPhoneId { get; set; }
+        public long AddressPhoneId { get; set; }
         public string Phone { get; set; }
-        public int AddressId { get; set; }
+        public long AddressId { get; set; }
 
     }
 }
