+++ Drs.Service/Track/ITrackService.cs
@@ -8,7 +8,7 @@ namespace Drs.Service.Track
     public interface ITrackService
     {
         IEnumerable<TrackOrderDto> SearchByPhone(PagerDto<String> phone);
-        IEnumerable<TrackOrderDto> SearchByClient(PagerDto<int> client);
+        IEnumerable<TrackOrderDto> SearchByClient(PagerDto<long> client);
         TrackOrderDto ShowDetailByOrderId(long orderId);
         IEnumerable<TrackOrderDto> SearchByDailyInfo(PagerDto<DailySearchModel> model);
     }
